+++ src/main/java/com/example/ilenguageapi/domain/service/SessionService.java
@@ -17,7 +17,7 @@ public interface SessionService {
     Session getSessionByEndAt(LocalDate endAt);
     Session getSessionByState(String state);
     Session getSessionByTopic(String topic);
-    Page<Session> getAllSessionsByScheduleId(int scheduleId, Pageable pageable);
-    Session getSessionByIdAndScheduleId(int scheduleId, Long sessionId);
+    Page<Session> getAllSessionsByScheduleId(Long scheduleId, Pageable pageable);
+    Session getSessionByIdAndScheduleId(Long scheduleId, Long sessionId);
    // ResponseEntity<?> deleteSession(int scheduleId, Long sessionId);
 }
